refactor: bestArticleCard 분리 (FE-42)
